Green: move matching logic into costcentre
